Recipe.md
# Black-Forest-Cake
ift together flour and cocoa powder in a large bowl.

Add sugar, baking soda and salt and mix well.

To make the buttermilk, mix 1 cup milk with white vinegar and set aside for 5 minutes.



After 5 minutes the milk will curdle a bit, at this point add oil and vanilla extract to it.

Add the dry ingredients to the wet and mix until the ingredients are just combined.

Add 2 tablespoons of water at this point if the batter seems too thick. This is optional, I did add it to my batter. Remember to not over-mix the batter.



Transfer batter to two 6 x 2 inch round pans lined with parchment paper.

Bake the cakes at 350 F degrees for 25-27 minutes or until a toothpick inserted in the center comes out clean.

While the cakes are in the oven, make the cherry syrup. To a pan on medium-high heat add water, sugar and cherries. You may add more sugar if using fresh cherries. These maraschino cherries are quite sweet so 2 tablespoons was enough. Let the syrup come to a boil and then let it simmer for 2-3 minutes.



Remove the syrup from heat and drain it into a bowl using a strainer. Transfer the cherries to another bowl and set aside.

Make the whipped cream. Place the whisk attachment and the steel bowl into the refrigerator for 20-30 minutes until chilled. Then add to it heavy cream and powdered sugar.

Using the wire whisk attachment of your stand mixer or using your hand mixer, beat the cream with powdered sugar at medium-high speed until it forms peaks.



To assemble the cake, first grate the chocolate or use a peeler to shave the chocolate. Use plain chocolate here.

Once the cakes have cooled down, cut each cake into 2 layer. You will total of 4 layers. Place the first layer on your serving plate or turn table. Brush the entire layer generously with the prepared cherry syrup.

Then apply whipped cream and then place some cherries on top. You may skip the cherries in layer, this is optional. I also cut cherries in half before adding to the layer.



Repeat again with another cake layer > brush with cherry syrup > whipped cream >cherries (if using). Repeat with the remaining layers.

Finally apply a thick layer of whipped cream on the top layer and also on the sides of the cake.

Pat down chocolate shavings on sides in the cake. Place the bigger chocolate curls in center.



Pipe the remaining whipped cream as rosettes. Top with remaining cherries and serve. I like to place this eggless black forest cake in the refrigerator for couple of hours before serving. Once it’s slightly set, its easier to cut into slices.